# src/main/resources/api.yaml
...
swagger: "2.0"
info:
  description: First API for the Tracey database models.
  version: 1.0.0
  title: Tracey_Database_API
  contact:
    email: fish_b@it.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /rikkygaray21/tracey-01/1.0.0
...
schemes:
- https
paths:
  /locations:
    get:
      summary: Get record of all locations
      operationId: getAllLocationRecords
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: Locations - Get all locations - success
        "400":
          description: Locations - Get all locations - failed
    post:
      summary: Create a location
      operationId: createLocationRecord
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: Locations - Post new location - success
        "400":
          description: Locations - Post new location - failed
  /locations/{locationId}:
    get:
      summary: Get a specific location by searching its locationId
      operationId: getSingleLocationById
      produces:
      - application/json
      parameters:
      - name: locationId
        in: path
        description: String ID of the location to get
        required: true
        type: string
      responses:
        "200":
          description: byLocationId - Get location by Id - success
        "400":
          description: byLocationId - Get location by Id - failed
    put:
      summary: Put/Modify a specific location by searching its locationId
      operationId: putSingleLocationById
      produces:
      - application/json
      parameters:
      - name: locationId
        in: path
        description: String ID of the location to put
        required: true
        type: string
      responses:
        "200":
          description: byLocationId - Put/modify location by Id - success
        "400":
          description: byLocationId - Put/modify location by Id - failed
    delete:
      summary: Delete a specific location by searching its locationId
      operationId: deleteSingleLocation
      produces:
      - application/json
      parameters:
      - name: locationId
        in: path
        description: Numeric ID of the location to delete
        required: true
        type: integer
      responses:
        "200":
          description: byLocationId - Delete location by Id - success
        "400":
          description: byLocationId - Delete location by Id - failed
    patch:
      summary: Patch/Modify a specific location by searching its locationId
      operationId: patchSingleLocationById
      produces:
      - application/json
      parameters:
      - name: locationId
        in: path
        description: String ID of the location to patch
        required: true
        type: string
      responses:
        "200":
          description: byLocationId - Patch/modify location by Id - success
        "400":
          description: byLocationId - Patch/modify location by Id - failed
  /locations/{locationId}/activities:
    get:
      summary: Gets all activities from a specific location.
      operationId: getActivitiesAtLocationById
      produces:
      - application/json
      parameters:
      - name: locationId
        in: path
        description: String ID of the location to get
        required: true
        type: string
      responses:
        "200":
          description: Activites_by_LocationId - Get activities by Id - success
        "400":
          description: Activites_by_LocationId - Get activities by Id - failed
    post:
      summary: Add an activity to a given location
      operationId: postActivityToLocationById
      produces:
      - application/json
      parameters:
      - name: locationId
        in: path
        description: String ID of the location to post
        required: true
        type: string
      responses:
        "200":
          description: Activity_by_LocationId - Post location by Id - success
        "400":
          description: Activity_by_LocationId - Post location by Id - failed
...